refactor: convert NatureSoundsApp to TypeScript
